+++ crates/pipeline-types/src/transport/url.rs
@@ -1,3 +1,4 @@
+use crate::config::TransportConfigVariant;
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 
@@ -7,4 +8,10 @@ use utoipa::ToSchema;
 pub struct UrlInputConfig {
     /// URL.
     pub path: String,
+}
+
+impl TransportConfigVariant for UrlInputConfig {
+    fn name(&self) -> String {
+        "url_input".to_string()
+    }
 }